requirements.txt
Django>=5.0
web3>=7.0.0
py-solc-x>=2.0.0
python-dotenv>=1.1.1
eth-account>=0.13.0
psycopg2-binary>=2.9.10
gunicorn>=23.0.0,<24.0.0
whitenoise>=6.0.0
gevent>=23.1.0
reportlab>=3.6.12
geopy>=2.2.0

dj-database-url>=1.0.0

